plugin cleanup on lifecycle hook
