+++ src/main/java/io/cryostat/discovery/DiscoveryPlugin.java
@@ -18,11 +18,13 @@ package io.cryostat.discovery;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.time.Instant;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.function.Supplier;
 
 import io.cryostat.credentials.Credential;
+import io.cryostat.discovery.KubeEndpointSlicesDiscovery.KubeDiscoveryNodeType;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -40,6 +42,7 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
+import jakarta.persistence.EntityManager;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
@@ -47,6 +50,7 @@ import jakarta.persistence.NamedQueries;
 import jakarta.persistence.NamedQuery;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreRemove;
 import jakarta.transaction.Transactional;
 import jakarta.validation.constraints.NotNull;
 import jakarta.ws.rs.GET;
@@ -147,11 +151,100 @@ public class DiscoveryPlugin extends PanacheEntityBase {
                 .singleResultOptional();
     }
 
+    /**
+     * Helper class for plugin cleanup operations in cases where nodes are not under the plugin's
+     * Realm, such as Agent instances using the publication fill algorithm.
+     */
+    @ApplicationScoped
+    static class PluginCleanupHelper {
+
+        @Inject EntityManager entityManager;
+        @Inject Logger logger;
+
+        void cleanupPluginNodes(DiscoveryPlugin plugin) {
+            cleanupPluginNodes(plugin.id);
+        }
+
+        void cleanupPluginNodes(UUID pluginId) {
+            // Clean up target nodes that belong to this plugin
+            // For KUBERNETES fill strategy, these are under KubernetesApi Realm, tagged with
+            // plugin ID
+            // For NONE fill strategy, cascade deletion handles cleanup when Agent Realm is
+            // deleted
+            List<DiscoveryNode> pluginNodes = DiscoveryNode.getByPluginId(pluginId);
+            if (pluginNodes == null || pluginNodes.isEmpty()) {
+                return;
+            }
+            for (DiscoveryNode node : pluginNodes) {
+                try {
+                    DiscoveryNode parent = node.parent;
+                    if (parent != null) {
+                        // Remove node from parent's collection before deleting
+                        // This ensures parent.hasChildren() will be accurate
+                        parent.children.remove(node);
+                        node.parent = null;
+                    }
+                    node.delete();
+                    // Recursively prune empty parent nodes up the tree
+                    pruneEmptyAncestors(parent);
+                } catch (Exception e) {
+                    logger.warn(e);
+                }
+            }
+        }
+
+        void pruneEmptyAncestors(DiscoveryNode child) {
+            // Walk up the hierarchy, removing childless nodes from their parents
+            // Follow the same pattern as KubeEndpointSlicesDiscovery.pruneOwnerChain()
+            while (child != null) {
+                DiscoveryNode parent = child.parent;
+
+                if (parent == null) {
+                    logger.debugv(
+                            "Reached orphaned node {0} (id={1}), relying on orphan removal",
+                            child.name, child.id);
+                    break;
+                }
+
+                entityManager.refresh(parent); // Reload from DB with current children
+
+                // Remove child from parent's collection - orphan removal will delete it
+                parent.children.remove(child);
+                child.parent = null;
+
+                boolean hasChildren = parent.hasChildren();
+                boolean isNamespace =
+                        parent.nodeType.equals(KubeDiscoveryNodeType.NAMESPACE.getKind());
+                boolean isRealm = parent.nodeType.equals("Realm");
+
+                logger.debugv(
+                        "Parent node {0} (id={1}, type={2}): hasChildren={3}, isNamespace={4},"
+                                + " isRealm={5}",
+                        parent.name, parent.id, parent.nodeType, hasChildren, isNamespace, isRealm);
+
+                // Stop pruning if:
+                // 1. Parent is a Realm (always preserve Realms)
+                // 2. Parent has children (preserve non-empty nodes)
+                // Note: Empty Namespaces should be pruned, so we don't stop just because it's a
+                // Namespace
+                if (isRealm || hasChildren) {
+                    logger.debugv("Stopping pruning at parent node {0}", parent.name);
+                    break;
+                }
+
+                logger.debugv("Continuing to prune parent node {0}", parent.name);
+                // Move up to check the parent too
+                child = parent;
+            }
+        }
+    }
+
     @ApplicationScoped
     static class Listener {
 
         @Inject Logger logger;
         @Inject PluginCallbackFactory callbackFactory;
+        @Inject PluginCleanupHelper cleanupHelper;
 
         @PrePersist
         @Transactional
@@ -189,6 +282,17 @@ public class DiscoveryPlugin extends PanacheEntityBase {
             } catch (Exception e) {
                 logger.error("Discovery Plugin ping failed", e);
                 throw e;
+            }
+        }
+
+        @PreRemove
+        @Transactional
+        public void preRemove(DiscoveryPlugin plugin) {
+            logger.debugv("PreRemove lifecycle hook triggered for plugin ID {0}", plugin.id);
+            try {
+                cleanupHelper.cleanupPluginNodes(plugin);
+            } catch (Exception e) {
+                logger.warn("Error during plugin cleanup", e);
             }
         }
